Cat-Dog-Classifications-using-CNN-master/app.py
# -*- coding: utf-8 -*-
"""
Created on Thu Jun 11 22:34:20 2020

@author: Shubham Gupta
"""

from __future__ import division, print_function
# coding=utf-8
import sys
import os
import glob
import re
import numpy as np

# Keras
from tensorflow.keras.applications.imagenet_utils import preprocess_input, decode_predictions
from tensorflow.keras.models import load_model
from tensorflow.keras.preprocessing import image
import tensorflow as tf
import cv2

# Flask utils
from flask import Flask, redirect, url_for, request, render_template
from werkzeug.utils import secure_filename
import numpy as np

# Define a flask app
app = Flask(__name__)

# Model saved with Keras model.save()
MODEL_PATH ='output_model.h5'

# Load your trained model
model = load_model(MODEL_PATH)


def model_predict(img_path):
    # predicting images
    img = image.load_img(img_path, target_size=(150, 150))
    x = image.img_to_array(img)
    x = np.expand_dims(x, axis=0)

    images = np.vstack([x])
    classes = model.predict(images, batch_size=10)
    print(classes[0])
    if classes[0] > 0.5:
        out = "This is a dog"
    else:
        out = "This is a cat"
    return out

@app.route('/', methods=['GET'])
def index():
    # Main page
    return render_template('index.html')


@app.route('/predict', methods=['GET', 'POST'])
def upload():
    if request.method == 'POST':
        # Get the file from post request
        f = request.files['file']

        # Save the file to ./uploads
        basepath = os.path.dirname(__file__)
        file_path = os.path.join(
            basepath, 'uploads', secure_filename(f.filename))
        f.save(file_path)

        # Make prediction
        prediction = model_predict(file_path)
        return prediction

    return None


if __name__ == '__main__':
    app.run(debug=True)
